Separated objects in directories except for system facade for a declarative importing. Changes object prices from ints to floats

+++ pkg/api/responses.go
@@ -23,7 +23,7 @@ type CheckOutCartResponse struct {
 	Message       string `json:"message,omitempty"`
 }
 type ListPurchasesResponse struct {
-	Status  int            `json:"status"`
-	Items   map[string]int `json:"items,omitempty"`
-	Message string         `json:"message,omitempty"`
+	Status  int                `json:"status"`
+	Items   map[string]float64 `json:"items,omitempty"`
+	Message string             `json:"message,omitempty"`
 }
